+++ jdaviz/configs/default/plugins/__init__.py
@@ -11,3 +11,4 @@ from .metadata_viewer.metadata_viewer import *  # noqa
 from .export_plot.export_plot import *  # noqa
 from .plot_options.plot_options import *  # noqa
 from .markers.markers import * # noqa
+from .histogram.histogram import *  # noqa