+++ jdaviz/configs/default/plugins/histogram/tests/test_histogram.py
@@ -1,0 +1,3 @@
+# TODO: Write tests!
+def test_histogram_imviz():
+    raise NotImplementedError